#319 moji.Token() has been moved to fact/lib.go
